+++ include/utils.sh
@@ -124,6 +124,17 @@ function copy_openams_config() {
   safe_copy "${afc_path}/config/macros" "${afc_config_dir}/"
 }
 
+function copy_snapmaker_config() {
+    mkdir -p "${afc_config_dir}"
+    safe_copy "${afc_path}/templates/u1_macros/AFC.cfg" "${afc_config_dir}/"
+    safe_copy "${afc_path}/config/AFC_Macro_Vars.cfg" "${afc_config_dir}/"
+    mkdir -p "${afc_config_dir}/mcu"
+    mkdir -p "${afc_config_dir}/macros"
+    safe_copy "${afc_path}/templates/u1_macros/Snapmaker_macros.cfg" "${afc_config_dir}/macros/"
+    safe_copy "${afc_path}/config/macros/AFC_macros.cfg" "${afc_config_dir}/AFC/macros"
+    safe_copy "${afc_path}/templates/AFC_Hardware_U1.cfg" "${afc_config_dir}/AFC_Hardware.cfg"
+}
+
 get_git_version() {
   local git_hash
   local afc_py_version
@@ -235,9 +246,18 @@ restart_service() {
 
 restart_klipper() {
   if query_printer_status; then
-    restart_service klipper
+    if [ "$is_snapmaker" == "True" ]; then
+      u1_restart_klipper
+    else
+      restart_service klipper
+    fi
+
   else
-    print_msg WARNING "Your printer is not idle/ready, or its status could not be confirmed. Automatic Klipper restart has been skipped; please restart the Klipper service manually once the printer is idle."
+    if [ "$is_snapmaker" == "True" ]; then
+      print_msg WARNING "Please manually restart klipper through <ip_address>/firmware-config panel, or fully power-cycle printer."
+    else
+      print_msg WARNING "Your printer is not idle/ready, or its status could not be confirmed. Automatic Klipper restart has been skipped; please restart the Klipper service manually once the printer is idle."
+    fi
   fi
 }
 
@@ -315,6 +335,10 @@ del_var_file() {
 check_for_k1() {
   if grep -Fqs "ID=buildroot" /etc/os-release; then
     is_k1_os="True"
+    if grep -Fqs RK_BUILD_INFO=\"snapmaker /etc/os-release; then
+      is_snapmaker="True"
+      klipper_venv="/usr/bin"
+    fi
   fi
 }